Second commit

+++ Beneficiaire-Service/src/main/java/org/enset/app/beneficiaireservice/services/BeneficiaireService.java
@@ -4,8 +4,10 @@ import org.enset.app.beneficiaireservice.entities.Beneficiaire;
 import org.enset.app.beneficiaireservice.repositories.BeneficiaireRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class BeneficiaireService {
@@ -25,7 +27,37 @@ public class BeneficiaireService {
         return beneficiaireRepository.save(beneficiaire);
     }
 
-    public void deleteBeneficiary(Long id) {
-        beneficiaireRepository.deleteById(id);
+    @Transactional
+    public Beneficiaire updateBeneficiary(Beneficiaire beneficiaire) {
+        // Vérifier si le bénéficiaire existe déjà
+        Optional<Beneficiaire> existingBeneficiary = beneficiaireRepository.findById(beneficiaire.getId());
+
+        if (existingBeneficiary.isPresent()) {
+            // Mettre à jour les champs nécessaires
+            Beneficiaire beneficiaryToUpdate = existingBeneficiary.get();
+            beneficiaryToUpdate.setFirstName(beneficiaire.getFirstName());
+            beneficiaryToUpdate.setLastName(beneficiaire.getLastName());
+            beneficiaryToUpdate.setType(beneficiaire.getType());
+            beneficiaryToUpdate.setRIB(beneficiaire.getRIB());
+            // Ajouter ici d'autres champs à mettre à jour
+
+            return beneficiaireRepository.save(beneficiaryToUpdate);
+        } else {
+            throw new IllegalArgumentException("Le bénéficiaire avec l'ID " + beneficiaire.getId() + " n'existe pas.");
+        }
     }
+
+    public void deleteBeneficiary(Long id) {
+        Optional<Beneficiaire> existingBeneficiary = beneficiaireRepository.findById(id);
+
+        if (existingBeneficiary.isPresent()) {
+            beneficiaireRepository.deleteById(id);
+        } else {
+            throw new IllegalArgumentException("Le bénéficiaire avec l'ID " + id + " n'existe pas.");
+        }
+    }
+
+//    public void deleteBeneficiary(Long id) {
+//        beneficiaireRepository.deleteById(id);
+//    }
 }
